+++ changelog.md
@@ -8,6 +8,7 @@ While all _Major_ releases should be reviewed, our only _large_ releases are **v
 
 <!-- #### 14.6.0 [Unreleased]
 - **[new]** - match term id
+- **[change]** - tag text by default on .concat('')
 -->
 
 #### 14.6.0 [Oct 2022]